+++ R/map.R
@@ -71,6 +71,21 @@
 #' To map over **columns** instead, first wrap a dataframe in [as.list()], or
 #' transpose a matrix using [t()].
 #'
+#' @section Nested Maps:
+#'
+#' At times you way wish to run maps within maps. To do this, the function
+#' provided to the outer map needs to include a call to [daemons()] to set
+#' daemons used by the inner map. To guard against inadvertently spawning an
+#' excessive number of daemons on the same machine, attempting to launch local
+#' daemons within a map using `daemons(n)` will error.
+#'
+#' A legitimate use of this pattern however is when the outer daemons are
+#' launched on remote machines, and you then wish to launch daemons locally on
+#' each of those machines. In this case, use the following solution: instead of
+#' a single call to `daemons(n)` make 2 separate calls to
+#' `daemons(url = local_url()); launch_local(n)`. This is equivalent, and is
+#' permitted from within a map.
+#'
 #' @examplesIf interactive()
 #' daemons(4)
 #'
@@ -146,9 +161,7 @@ mirai_map <- function(
   .promise = NULL,
   .compute = NULL
 ) {
-  if (is.null(.compute)) .compute <- .[["cp"]]
-  envir <- ..[[.compute]]
-  is.null(envir) && stop(._[["requires_daemons"]])
+  require_daemons(call = environment(), .compute = .compute)
   is.function(.f) || stop(sprintf(._[["function_required"]], typeof(.f)))
 
   dx <- dim(.x)
@@ -160,7 +173,7 @@ mirai_map <- function(
           mirai(
             .expr = do.call(.f, c(list(.x), .args), quote = TRUE),
             ...,
-            .args = list(.f = .f, .x = x, .args = .args),
+            .args = list(.f = .f, .x = x, .args = .args, .mirai_within_map = TRUE),
             .compute = .compute
           )
       ),
@@ -175,7 +188,7 @@ mirai_map <- function(
             mirai(
               .expr = do.call(.f, c(as.vector(.x, mode = "list"), .args), quote = TRUE),
               ...,
-              .args = list(.f = .f, .x = .x[i, ], .args = .args),
+              .args = list(.f = .f, .x = .x[i, ], .args = .args, .mirai_within_map = TRUE),
               .compute = .compute
             )
         ),
@@ -190,7 +203,7 @@ mirai_map <- function(
             mirai(
               .expr = do.call(.f, c(.x, .args), quote = TRUE),
               ...,
-              .args = list(.f = .f, .x = lapply(.x, `[[`, i), .args = .args),
+              .args = list(.f = .f, .x = lapply(.x, `[[`, i), .args = .args, .mirai_within_map = TRUE),
               .compute = .compute
             )
         ),
@@ -212,19 +225,15 @@ mirai_map <- function(
 `[.mirai_map` <- function(x, ...) {
   missing(..1) && return(collect_aio_(x))
 
-  ensure_cli_initialized()
   dots <- eval(`[[<-`(substitute(alist(...)), 1L, quote(list)), envir = .)
-  map(x, dots)
+  mmap(x, dots)
 }
 
 #' @export
 #'
 print.mirai_map <- function(x, ...) {
   xlen <- length(x)
-  cat(
-    sprintf("< mirai map [%d/%d] >\n", xlen - .unresolved(x), xlen),
-    file = stdout()
-  )
+  cat(sprintf("< mirai map [%d/%d] >\n", xlen - .unresolved(x), xlen), file = stdout())
   invisible(x)
 }
 
@@ -244,10 +253,7 @@ print.mirai_map <- function(x, ...) {
     ) {
       is_error_value(xi) && {
         stop_mirai(x)
-        stop(
-          sprintf("In index %d:\n%s", i, attr(xi, "message")),
-          call. = FALSE
-        )
+        stop(sprintf("In index %d:\n%s", i, attr(xi, "message")), call. = FALSE)
       }
       typeof(xi) != typ && {
         stop_mirai(x)
@@ -294,21 +300,16 @@ print.mirai_map <- function(x, ...) {
 
 # internals --------------------------------------------------------------------
 
-ensure_cli_initialized <- function()
-  if (is.null(.[[".flat"]])) {
-    cli <- requireNamespace("cli", quietly = TRUE)
-    `[[<-`(
-      `[[<-`(
-        `[[<-`(., ".flat", if (cli) flat_cli else .flat),
-        ".progress",
-        if (cli) progress_cli else .progress
-      ),
-      ".stop",
-      if (cli) stop_cli else .stop
-    )
-  }
+ensure_cli_initialized <- function() {
+  is.null(.[["require_daemons"]]) || return()
+  cli <- requireNamespace("cli", quietly = TRUE)
+  `[[<-`(., ".flat", if (cli) flat_cli else .flat)
+  `[[<-`(., ".progress", if (cli) progress_cli else .progress)
+  `[[<-`(., ".stop", if (cli) stop_cli else .stop)
+  `[[<-`(., "require_daemons", if (cli) stop_d_cli else stop_d)
+}
 
-map <- function(x, dots) {
+mmap <- function(x, dots) {
   expr <- if (length(dots) > 1L) do.call(expression, dots) else dots[[1L]]
   xlen <- length(x)
   i <- 0L