+++ R/next.R
@@ -26,7 +26,7 @@
 #' nextstream()
 #'
 #' nextget("pid")
-#' nextget("urls")
+#' nextget("url")
 #'
 #' daemons(0)
 #'
@@ -40,8 +40,9 @@ nextstream <- function(.compute = "default") next_stream(..[[.compute]])
 #' `nextget` retrieves the specified item from the specified compute profile.
 #'
 #' @param x character value of item to retrieve. One of `"n"` (number of
-#'   dispatcher daemons), `"pid"` (dispatcher process ID), `"urls"` (URLs
-#'   dispatcher is listening at) or `"tls"` (the stored client TLS configuration
+#'   dispatcher daemons), `"pid"` (dispatcher process ID), `"dispatcher"` (the
+#'   URL to connect to dispatcher from the host) `"url"` (the URL to connect to
+#'   dispatcher from daemons) or `"tls"` (the stored client TLS configuration
 #'   for use by daemons).
 #'
 #' @return For `nextget`: the requested item, or else NULL if not present.
@@ -87,10 +88,4 @@ next_stream <- function(envir) {
   stream <- envir[["stream"]]
   if (is.integer(stream)) `[[<-`(envir, "stream", parallel::nextRNGStream(stream))
   stream
-}
-
-next_msgid <- function(envir) {
-  prev <- envir[["msgid"]]
-  is.null(prev) && return()
-  envir[["msgid"]] <- prev + 1L
 }